Revert "Merge pull request #94 from VisualTiles/SetChannelQrCodeAsAService"
This reverts commit 0a01b93b937275197364ed7b8d96e145748fe6fc, reversing
changes made to 9712253343108a5721f1ccb718be3fd7303f2a76.
