+++ app/src/main/java/com/javierarboleda/visualtilestogether/activities/TileListActivity.java
@@ -137,7 +137,7 @@ public class TileListActivity extends AppCompatActivity implements GoogleApiClie
                 startActivity(new Intent(this, ModeratorConsoleActivity.class));
                 break;
             case R.id.nav_normalize_db:
-                normalizeDb(this);
+                normalizeDb();
                 break;
             case R.id.nav_show_code:
                 startActivity(new Intent(this, ShowQrCodeActivity.class));
