+++ app/src/main/java/com/javierarboleda/visualtilestogether/fragments/TileListFragmentInChannel.java
@@ -11,7 +11,7 @@ public class TileListFragmentInChannel extends TileListFragment {
 
     Query getDbQuery(DatabaseReference dbRef) {
         return dbRef
-                .child(visualTilesTogetherApp.getChannelId())
+                .child(visualTilesTogetherApp.getUser().getChannelId())
                 .child(Channel.TILE_IDS)
                 .orderByKey();
     }
